requirements.txt
Django==4.0.5
django-filter==22.1
djangorestframework==3.13.1
drf-yasg==1.20.0
numpy==1.23.0
openpyxl==3.0.10
pandas==1.4.3
python-dateutil==2.8.2
python-decouple==3.6
rest-pandas==1.1.0
